{"schema":"decodex/commit/1","summary":"Split account auth JSON","authority":"XY-1115"}
